requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pillow
python-multipart
pytest
httpx
